FC-CNN with SE and class_weights
The class weights have been calculated by counting how many times each class appears in the dataset (and not counting the number of pixels of each class). For example, if in an immage there is only class 1 (in all the pixels), the counter increments by one. It doesn't work so well, I don't know if it really does what I thought.